+++ lib/src/dpgsql_db_type.dart
@@ -26,6 +26,7 @@ enum DpgsqlDbType {
   timestamp,
   timestampTz,
   uuid,
+  bit,
   varbit,
   varchar,
   xml,